# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LoanGo API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface UserProfile {
  userId: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  dob?: string | null;
  /** @nullable */
  pan?: string | null;
  /** @nullable */
  employmentType?: string | null;
  /** @nullable */
  monthlyIncome?: number | null;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  workExperienceMonths?: number | null;
  /** @nullable */
  addressLine?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  pincode?: string | null;
  acceptedTerms: boolean;
  acceptedPrivacy: boolean;
  kycCompleted: boolean;
}

export interface ProfileUpdate {
  phone?: string;
  dob?: string;
  pan?: string;
  employmentType?: string;
  /** @minimum 0 */
  monthlyIncome?: number;
  companyName?: string;
  /** @minimum 0 */
  workExperienceMonths?: number;
  addressLine?: string;
  city?: string;
  state?: string;
  pincode?: string;
  acceptedTerms?: boolean;
  acceptedPrivacy?: boolean;
}

export interface LoanOffer {
  id: string;
  lender: string;
  productName: string;
  loanType: string;
  /** @nullable */
  description?: string | null;
  minAmount: number;
  maxAmount: number;
  interestMin: number;
  interestMax: number;
  tenureMinMonths: number;
  tenureMaxMonths: number;
  processingFeePct: number;
  rating: number;
  featured: boolean;
}

export type LoanApplicationStatus = typeof LoanApplicationStatus[keyof typeof LoanApplicationStatus];


export const LoanApplicationStatus = {
  submitted: 'submitted',
  under_review: 'under_review',
  approved: 'approved',
  rejected: 'rejected',
  disbursed: 'disbursed',
} as const;

export interface LoanApplication {
  id: string;
  userId: string;
  offerId: string;
  amount: number;
  tenureMonths: number;
  purpose: string;
  status: LoanApplicationStatus;
  /** @nullable */
  rejectionReason?: string | null;
  createdAt: string;
}

export interface LoanApplicationInput {
  /** @minLength 1 */
  offerId: string;
  /** @minimum 1 */
  amount: number;
  /** @minimum 1 */
  tenureMonths: number;
  /** @minLength 1 */
  purpose: string;
}

export type LoanStatus = typeof LoanStatus[keyof typeof LoanStatus];


export const LoanStatus = {
  active: 'active',
  closed: 'closed',
  defaulted: 'defaulted',
} as const;

export interface Loan {
  id: string;
  applicationId: string;
  userId: string;
  principal: number;
  interestRate: number;
  tenureMonths: number;
  emiAmount: number;
  totalPayable: number;
  amountPaid: number;
  status: LoanStatus;
  disbursedAt: string;
  /** @nullable */
  nextDueDate?: string | null;
}

export type EmiPaymentStatus = typeof EmiPaymentStatus[keyof typeof EmiPaymentStatus];


export const EmiPaymentStatus = {
  pending: 'pending',
  paid: 'paid',
  overdue: 'overdue',
} as const;

export interface EmiPayment {
  id: string;
  loanId: string;
  installmentNumber: number;
  dueDate: string;
  amount: number;
  status: EmiPaymentStatus;
  /** @nullable */
  paidAt?: string | null;
}

export interface LoanDetail {
  loan: Loan;
  emis: EmiPayment[];
}

export interface CheckoutSession {
  url: string;
  sessionId?: string;
}

export interface Notification {
  id: string;
  userId: string;
  type: string;
  title: string;
  body: string;
  read: boolean;
  createdAt: string;
}

export type KycDocumentStatus = typeof KycDocumentStatus[keyof typeof KycDocumentStatus];


export const KycDocumentStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface KycDocument {
  id: string;
  userId: string;
  docType: string;
  objectKey: string;
  status: KycDocumentStatus;
  uploadedAt: string;
}

export interface KycDocumentInput {
  /** @minLength 1 */
  docType: string;
  /** @minLength 1 */
  objectKey: string;
}

export interface KycUploadRequest {
  /** @minLength 1 */
  docType: string;
  /** @minLength 1 */
  contentType: string;
}

export interface KycUploadResponse {
  uploadUrl: string;
  objectKey: string;
}

/**
 * Error
 */
export type ErrorResponseResponse = ErrorEnvelope;

export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};