# lib/api-zod/src/generated/types/emiPayment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LoanGo API
 * OpenAPI spec version: 0.1.0
 */
import type { EmiPaymentStatus } from './emiPaymentStatus';

export interface EmiPayment {
  id: string;
  loanId: string;
  installmentNumber: number;
  dueDate: Date;
  amount: number;
  status: EmiPaymentStatus;
  /** @nullable */
  paidAt?: Date | null;
}