# lib/api-zod/src/generated/types/kycDocument.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LoanGo API
 * OpenAPI spec version: 0.1.0
 */
import type { KycDocumentStatus } from './kycDocumentStatus';

export interface KycDocument {
  id: string;
  userId: string;
  docType: string;
  objectKey: string;
  status: KycDocumentStatus;
  uploadedAt: Date;
}