# lib/api-zod/src/generated/types/kycDocumentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LoanGo API
 * OpenAPI spec version: 0.1.0
 */

export interface KycDocumentInput {
  /** @minLength 1 */
  docType: string;
  /** @minLength 1 */
  objectKey: string;
}