# lib/api-zod/src/generated/types/kycUploadResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LoanGo API
 * OpenAPI spec version: 0.1.0
 */

export interface KycUploadResponse {
  uploadUrl: string;
  objectKey: string;
}