# lib/api-zod/src/generated/types/loanApplication.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LoanGo API
 * OpenAPI spec version: 0.1.0
 */
import type { LoanApplicationStatus } from './loanApplicationStatus';

export interface LoanApplication {
  id: string;
  userId: string;
  offerId: string;
  amount: number;
  tenureMonths: number;
  purpose: string;
  status: LoanApplicationStatus;
  /** @nullable */
  rejectionReason?: string | null;
  createdAt: Date;
}